+++ src/libs/PolicyEmployeeListUtils.ts
@@ -2,7 +2,6 @@ import type {OnyxCollection} from 'react-native-onyx';
 import Onyx from 'react-native-onyx';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {Policy} from '@src/types/onyx';
-import {getCurrentUserAccountID} from './actions/Report';
 import {getPolicyEmployeeListByIdWithoutCurrentUser} from './PolicyUtils';
 
 let allPolicies: OnyxCollection<Policy> = {};
@@ -12,12 +11,22 @@ Onyx.connect({
     callback: (value) => (allPolicies = value),
 });
 
+let currentUserAccountID: number | undefined;
+Onyx.connect({
+    key: ONYXKEYS.SESSION,
+    callback: (value) => {
+        currentUserAccountID = value?.accountID;
+    },
+});
+
 function getPolicyEmployeeAccountIDs(policyID?: string) {
     if (!policyID) {
         return [];
     }
 
-    const currentUserAccountID = getCurrentUserAccountID();
+    if (!currentUserAccountID) {
+        return [];
+    }
 
     return getPolicyEmployeeListByIdWithoutCurrentUser(allPolicies, policyID, currentUserAccountID);
 }
